Reimplement preprints as a set of fields on a node...
...rather than a separate model
