+++ osf_models/models/__init__.py
@@ -23,5 +23,4 @@ from osf_models.models.watch_config import WatchConfig  # noqa
 from osf_models.models.wiki import NodeWikiPage  # noqa
 from osf_models.models.subject import Subject  # noqa
 from osf_models.models.preprint_provider import PreprintProvider  # noqa
-from osf_models.models.preprint import Preprint  # noqa
 from osf_models.models.identifiers import Identifier  # noqa
